model/requirements.txt
requests~=2.32.5
numpy~=2.3.5
pillow~=12.0.0
openai
rdflib
